Java

+++ src/main/java/com/mycompany/web/jpa/controller/UserServlet.java
@@ -1,10 +1,8 @@
 package com.mycompany.web.jpa.controller;
 
 import com.google.gson.Gson;
-import com.mycompany.web.jpa.entity.User;
 import com.mycompany.web.jpa.service.UserService;
 import java.io.IOException;
-import java.util.Random;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -18,7 +16,7 @@ public class UserServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         UserService service = new UserService();
-        long user_id, user_age;
+        Integer user_id, user_age;
         String user_name;
         
         String pathInfo = req.getPathInfo();
@@ -26,17 +24,17 @@ public class UserServlet extends HttpServlet {
         switch(pathInfo) {
             case "/insert":
                 user_name = req.getParameter("user_name");
-                user_age = Long.parseLong(req.getParameter("user_age"));
+                user_age = Integer.parseInt(req.getParameter("user_age"));
                 resp.getWriter().print(service.insert(user_name, user_age));
                 return;
             case "/update":
-                user_id = Long.parseLong(req.getParameter("user_id"));
+                user_id = Integer.parseInt(req.getParameter("user_id"));
                 user_name = req.getParameter("user_name");
-                user_age = Long.parseLong(req.getParameter("user_age"));
+                user_age = Integer.parseInt(req.getParameter("user_age"));
                 resp.getWriter().print(service.update(user_id, user_name, user_age));
                 return;
             case "/delete":
-                user_id = Long.parseLong(req.getParameter("user_id"));
+                user_id = Integer.parseInt(req.getParameter("user_id"));
                 resp.getWriter().print(service.delete(user_id));
                 return;
         }
