+++ src/main/java/com/mycompany/web/jpa/service/UserService.java
@@ -1,6 +1,5 @@
 package com.mycompany.web.jpa.service;
 
-import com.mycompany.web.jpa.dao.CustomerDAO;
 import com.mycompany.web.jpa.entity.User;
 import com.mycompany.web.jpa.dao.UserDAO;
 import com.mycompany.web.jpa.util.JPAUtil;
@@ -16,22 +15,26 @@ public class UserService {
         dao = new UserDAO(em);
     }
     
-    public boolean insert(String user_name, long user_age) {
+    // 新增
+    public boolean insert(String user_name, Integer user_age) {
         User user = new User();
         user.setName(user_name);
         user.setAge(user_age);
         return dao.insert(user);
     }
     
+    // 查詢多筆
     public List<User> query() {
         return dao.query();
     }
     
-    public User findById(Long id) {
+    // 查詢單筆
+    public User findById(Integer id) {
         return dao.findById(id);
     }
     
-    public boolean update(Long id, String name, Long age) {
+    // 修改
+    public boolean update(Integer id, String name, Integer age) {
         User user = findById(id);
         if(user == null) {
             return false;
@@ -42,7 +45,8 @@ public class UserService {
         }
     }
     
-    public boolean delete(Long id) {
+    // 刪除
+    public boolean delete(Integer id) {
         User user = findById(id);
         return (user == null) ? false : dao.delete(user);
     }
